refactor(frontend): two record shells move to the design system, the third is gone
frontend/AGENTS.md: a screen file is not a place to keep a primitive, and if a
screen exports something another screen imports it belongs in
src/design-system/. record360/shells.tsx exported three components, two of
them already imported from outside record360/.

RailPanel joins panel.tsx, the shape it composes: it is a Panel that knows the
difference between "there is nothing here" and "you may not read this".
deal360/dealseats.tsx now reaches it directly.

OverlayFallback gets overlayfallback.tsx of its own — it replaces a whole page
rather than filling a panel. organizations.tsx now reaches it directly.

SectionCard is DELETED rather than moved. #3867 replaced its call sites with
RailPanel yesterday and nothing has drawn it since; only its own stories
referenced it. Promoting an uncalled component into the catalog would put a
primitive there that no page can show a reader.

That also dissolves the obstacle #3903 describes. SectionCard was the only one
of the three that rendered `.co-card`, so shells.tsx no longer needs to import
a screen's stylesheet, and the rule stays in company360.css where its other
callers — organizations.tsx draws `<Card className="co-card">` twice — already
put it.

incompleteGraph is a predicate rather than a component and moves to
graphcompleteness.ts, which leaves no module in record360/ reaching into a
screen stylesheet.
